% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStockEMA.R
\name{calculateEma}
\alias{calculateEma}
\title{Calculates an exponential moving average (EMA) for a provided stock}
\usage{
calculateEma(x, n = 10)
}
\arguments{
\item{x}{A numeric vector or xts object containing price data}

\item{n}{The number of periods to use for the EMA calculation}
}
\value{
An xts object or numeric vector containing the EMA values
}
\description{
Computes an EMA for a given numeric or xts (a time‑indexed data structure) price series.
}
