% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeeklyStockPrice.R
\name{getStockWeekly}
\alias{getStockWeekly}
\title{Gets the weekly data of a stock using the OHLC data}
\usage{
getStockWeekly(ticker)
}
\arguments{
\item{ticker}{A stock ticker symbol (for example, "AAPL", or "IBM")}
}
\value{
Weekly OHLC data as an xts object
}
\description{
Downloads weekly OHLC data for a given stock ticker.
}
