% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortTermTrendComparison.R
\name{shortTermSignal}
\alias{shortTermSignal}
\title{Generates a short-term trading signal based on EMA crossovers, that is if the 10 week is above the 20, its bullish, opposite being true means bearish}
\usage{
shortTermSignal(ticker)
}
\arguments{
\item{ticker}{A stock ticker symbol (for example, "AAPL", "IBM")}
}
\value{
A sentence stating that the stock is in a bullish trend if the 10-week EMA is above the 20-week EMA,
otherwise it is in a bearish trend, 20 above the 10.
}
\description{
Uses the 10-week and 20-week EMAs to determine whether a stock is in a
short-term bullish or bearish trend.
}
